Convert ergm coefficients to fold changes

The fold changes are all based on comparison to reference level

Usage

get_ergm_fc(ergm, trim = TRUE)

Arguments

ergm

an ergm fit

trim

whether to remove coefficients where the coefficient is -Inf

Value

tibble of fold changes

Examples

ergm <- get_ergms(example_network, preds = "site")
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.
#> Evaluating log-likelihood at the estimate. 
#> 
#> Starting maximum pseudolikelihood estimation (MPLE):
#> Obtaining the responsible dyads.
#> Evaluating the predictor and response matrix.
#> Maximizing the pseudolikelihood.
#> Finished MPLE.
#> Evaluating log-likelihood at the estimate. 
#> 
get_ergm_fc(ergm[[1]])
#> # A tibble: 6 × 3
#>   term          theta   phi
#>   <chr>         <dbl> <dbl>
#> 1 edges        -1.62  1    
#> 2 mix.site.A.B -1.37  0.289
#> 3 mix.site.B.B  0.642 1.65 
#> 4 mix.site.A.C -0.868 0.464
#> 5 mix.site.B.C -1.25  0.325
#> 6 mix.site.C.C  1.64  3.07